services/applications/block/executor_proxy/requirements.txt
grpcio==1.62.1
protobuf==3.20.3
redis==3.2.1
